Create folder for mvvm
Create different folders to work according to the MVVM architecture. Model folder is created to store the classes and data structures that represent the application's data and business logic. View folder is created to store the user interface components, such as XAML files or code-behind files, responsible for displaying the application's visual elements to the user. ViewModel folder is created to store the classes that act as intermediate between the model and the view in the MVVM architecture. Services folder is created to store the classes that provide additional functionality to the application, such as data access, web services, or other external integration