Add color to resource folder
This commit adds color resources with opacity to the application. The BorderColor resource now includes an opacity value (#BCBCBC80) to achieve a partially transparent effect. This allows for more flexible and visually appealing UI designs by providing control over the opacity of certain elements. The new color resources, such as AppColor, PlaceholderColor, and GreyColor, can be easily referenced and applied throughout the application to enhance the user interface.