Create Directory List Screen
The XAML code you provided represents a ContentPage with a ListView to display a list of directories. Each item in the list is displayed using a ViewCell with a Grid layout. The Name property of the AlumniModel is bound to a Label to display the name of the directory. Additionally, an Image is used to display an icon.
The code snippet you provided looks correct for displaying the list of directories in a ListView. The ItemsSource property is bound to the AluminiList property in the view model, and the ItemTemplate defines the layout for each item in the list. The Label binds to the Name property of the AlumniModel, and the Image displays the icon using the specified source image file.